update to zig 0.13.0 and simplify build.zig
Consolidated the sample build code into one top-level `build.zig`.
This deduplicates code and will make it easier to provide the
bindings as a zig package in the future.

Also, instead of compiling the `liborca_wasm.a` module from scratch
we link it via the sdk path found by calling `orca sdk-path`.